done with the Sadegh task list

+++ app/api/task/delete.php
@@ -11,6 +11,18 @@ include_once "../../core/initialize.php";
 
 
 use App\Core\task;
+use Firebase\JWT\JWT;
+use Firebase\JWT\Key;
+require_once('../../../vendor/autoload.php');
+$jwt = new JWT;
+$headers = apache_request_headers();
+foreach ($headers as $header => $value) {
+    if ($header === "Authorization") {
+        $token = substr($value, 7);
+    }
+}
+$json = (array) $jwt->decode($token, new Key("token", 'HS512'));
+
 
 $task = new task($db);
 
@@ -19,6 +31,10 @@ $data = json_decode(file_get_contents("php://input"));
 
 $task->id = $data->id;
 
+if (!$task->user_name === $json['username']) {
+    echo json_encode(array("message" => "no auth", "username" => $task->user_name["user_name"]));
+    exit;
+}
 
 
 if ($task->delete()) {
